+++ src/js/modal.js
@@ -3,7 +3,7 @@
 
   var defaults;
   
-  $.modal = function(params) {
+  $.modal = function(params, onOpen) {
     params = $.extend({}, defaults, params);
 
 
@@ -19,7 +19,7 @@
                 '<div class="weui_dialog_ft">' + buttonsHtml + '</div>' +
               '</div>';
     
-    var dialog = $.openModal(tpl);
+    var dialog = $.openModal(tpl, onOpen);
 
     dialog.find(".weui_btn_dialog").each(function(i, e) {
       var el = $(e);
@@ -28,21 +28,30 @@
         if(params.autoClose) $.closeModal();
 
         if(buttons[i].onClick) {
-          buttons[i].onClick();
+          buttons[i].onClick.call(dialog);
         }
       });
     });
+
+    return dialog;
   };
 
-  $.openModal = function(tpl) {
+  $.openModal = function(tpl, onOpen) {
     var mask = $("<div class='weui_mask'></div>").appendTo(document.body);
     mask.show();
 
     var dialog = $(tpl).appendTo(document.body);
-    
+ 
+    if (onOpen) {
+      dialog.transitionEnd(function () {
+        onOpen.call(dialog);
+      });
+    }   
+
     dialog.show();
     mask.addClass("weui_mask_visible");
     dialog.addClass("weui_dialog_visible");
+
 
     return dialog;
   }
@@ -56,69 +65,183 @@
     });
   };
 
-  $.alert = function(text, title, callback) {
-    if (typeof title === 'function') {
-      callback = arguments[1];
-      title = undefined;
+  $.alert = function(text, title, onOK) {
+    var config;
+    if (typeof text === 'object') {
+      config = text;
+    } else {
+      if (typeof title === 'function') {
+        onOK = arguments[1];
+        title = undefined;
+      }
+
+      config = {
+        text: text,
+        title: title,
+        onOK: onOK
+      }
     }
     return $.modal({
-      text: text,
-      title: title,
+      text: config.text,
+      title: config.title,
       buttons: [{
         text: defaults.buttonOK,
         className: "primary",
-        onClick: callback
+        onClick: config.onOK
       }]
     });
   }
 
-  $.confirm = function(text, title, callbackOK, callbackCancel) {
-    if (typeof title === 'function') {
-      callbackCancel = arguments[2];
-      callbackOK = arguments[1];
-      title = undefined;
+  $.confirm = function(text, title, onOK, onCancel) {
+    var config;
+    if (typeof text === 'object') {
+      config = text
+    } else {
+      if (typeof title === 'function') {
+        onCancel = arguments[2];
+        onOK = arguments[1];
+        title = undefined;
+      }
+
+      config = {
+        text: text,
+        title: title,
+        onOK: onOK,
+        onCancel: onCancel
+      }
     }
     return $.modal({
-      text: text,
-      title: title,
+      text: config.text,
+      title: config.title,
       buttons: [
       {
         text: defaults.buttonCancel,
         className: "default",
-        onClick: callbackCancel
+        onClick: config.onCancel
       },
       {
         text: defaults.buttonOK,
         className: "primary",
-        onClick: callbackOK
+        onClick: config.onOK
       }]
     });
   };
 
-  $.prompt = function(text, title, callbackOK, callbackCancel) {
-    if (typeof title === 'function') {
-      callbackCancel = arguments[2];
-      callbackOK = arguments[1];
-      title = undefined;
+  //如果参数过多，建议通过 config 对象进行配置，而不是传入多个参数。
+  $.prompt = function(text, title, onOK, onCancel, input) {
+    var config;
+    if (typeof text === 'object') {
+      config = text;
+    } else {
+      if (typeof title === 'function') {
+        input = arguments[3];
+        onCancel = arguments[2];
+        onOK = arguments[1];
+        title = undefined;
+      }
+      config = {
+        text: text,
+        title: title,
+        input: input,
+        onOK: onOK,
+        onCancel: onCancel,
+        empty: false  //allow empty
+      }
     }
 
-    return $.modal({
-      text: "<p class='weui-prompt-text'>"+(text || "")+"</p><input type='text' class='weui_input weui-prompt-input' id='weui-prompt-input'/>",
-      title: title,
+    var modal = $.modal({
+      text: '<p class="weui-prompt-text">'+(config.text || '')+'</p><input type="text" class="weui_input weui-prompt-input" id="weui-prompt-input" value="' + (config.input || '') + '" />',
+      title: config.title,
+      autoClose: false,
       buttons: [
       {
         text: defaults.buttonCancel,
         className: "default",
-        onClick: callbackCancel
+        onClick: function () {
+          $.closeModal();
+          config.onCancel && config.onCancel.call(modal);
+        }
       },
       {
         text: defaults.buttonOK,
         className: "primary",
         onClick: function() {
-          callbackOK && callbackOK($("#weui-prompt-input").val());
+          var input = $("#weui-prompt-input").val();
+          if (!config.empty && (input === "" || input === null)) {
+            modal.find('.weui-prompt-input').focus()[0].select();
+            return false;
+          }
+          $.closeModal();
+          config.onOK && config.onOK.call(modal, input);
         }
       }]
+    }, function () {
+      this.find('.weui-prompt-input').focus()[0].select();
     });
+
+    return modal;
+  };
+
+  //如果参数过多，建议通过 config 对象进行配置，而不是传入多个参数。
+  $.login = function(text, title, onOK, onCancel, username, password) {
+    var config;
+    if (typeof text === 'object') {
+      config = text;
+    } else {
+      if (typeof title === 'function') {
+        password = arguments[4];
+        username = arguments[3];
+        onCancel = arguments[2];
+        onOK = arguments[1];
+        title = undefined;
+      }
+      config = {
+        text: text,
+        title: title,
+        username: username,
+        password: password,
+        onOK: onOK,
+        onCancel: onCancel
+      }
+    }
+
+    var modal = $.modal({
+      text: '<p class="weui-prompt-text">'+(config.text || '')+'</p>' +
+            '<input type="text" class="weui_input weui-prompt-input" id="weui-prompt-username" value="' + (config.username || '') + '" placeholder="输入用户名" />' +
+            '<input type="password" class="weui_input weui-prompt-input" id="weui-prompt-password" value="' + (config.password || '') + '" placeholder="输入密码" />',
+      title: config.title,
+      autoClose: false,
+      buttons: [
+      {
+        text: defaults.buttonCancel,
+        className: "default",
+        onClick: function () {
+          $.closeModal();
+          config.onCancel && config.onCancel.call(modal);
+        }
+      }, {
+        text: defaults.buttonOK,
+        className: "primary",
+        onClick: function() {
+          var username = $("#weui-prompt-username").val();
+          var password = $("#weui-prompt-password").val();
+          if (!config.empty && (username === "" || username === null)) {
+            modal.find('#weui-prompt-username').focus()[0].select();
+            return false;
+          }
+          if (!config.empty && (password === "" || password === null)) {
+            modal.find('#weui-prompt-password').focus()[0].select();
+            return false;
+          }
+          $.closeModal();
+          config.onOK && config.onOK.call(modal, username, password);
+        }
+      }]
+    }, function () {
+      this.find('#weui-prompt-username').focus()[0].select();
+    });
+
+    return modal;
   };
 
   defaults = $.modal.prototype.defaults = {